group pkg, move some stray files
