+++ core/src/main/scala/hammerlab/iterator/package.scala
@@ -4,6 +4,7 @@ package object iterator
   extends bulk
      with count
      with drop
+     with group
      with scan
      with sliding
      with sorted
