+++ core/src/main/scala/org/hammerlab/iterator/Bulk.scala
@@ -1,8 +1,11 @@
 package org.hammerlab.iterator
 
+import scala.collection.mutable
+import scala.reflect.ClassTag
+
 trait Bulk {
   /**
-   * Some smarter bulk operations on [[BufferedIterator]]s
+   * Some smarter bulk operations on [[BufferedIterator]]s, consuming exactly the elements necessary and not more
    */
   implicit class BufferedBulkOps[T](it: BufferedIterator[T]) extends Serializable {
     def takewhile(fn: T ⇒ Boolean): SimpleBufferedIterator[T] =
@@ -30,5 +33,26 @@ trait Bulk {
           else
             None
       }
+  }
+
+  /**
+   * Exposes `.takeEager` on iterators, which is identical to `.take.toArray`, but always consumes the taken elements from
+   * the original iterator.
+   *
+   * Some stdlib iterators backed by [[Seq]]s have `.take` implementations that create a view over the "taken" elements,
+   * so that the original iterator is not consumed when the taken/prefix iterator is materialized; this homogenizes that
+   * behavior.
+   */
+  implicit class TakeEagerIterator[T: ClassTag](it: Iterator[T]) {
+    def takeEager(n: Int): Array[T] = {
+      val builder = mutable.ArrayBuilder.make[T]()
+      for {
+        _ ← 0 until n
+        if it.hasNext
+      } {
+        builder += it.next
+      }
+      builder.result
+    }
   }
 }
