+++ core/src/test/scala/org/hammerlab/iterator/group/RunLengthTest.scala
@@ -1,13 +1,13 @@
-package org.hammerlab.iterator
+package org.hammerlab.iterator.group
 
-import org.hammerlab.iterator.RunLengthIterator._
+import hammerlab.iterator.group._
 import org.hammerlab.test.Suite
 
-class RunLengthIteratorTest extends Suite {
+class RunLengthTest extends Suite {
 
   {
     def check[T](elems: T*)(expected: (T, Int)*): Unit = {
-      elems.iterator.runLengthEncode().toSeq should be(expected)
+      elems.runLengthEncode().toSeq should be(expected)
     }
 
     test("empty") {
@@ -48,7 +48,7 @@ class RunLengthIteratorTest extends Suite {
       Ordering.by[Int, Int](_ % 2)
 
     def check(elems: Int*)(expected: (Int, Int)*): Unit = {
-      elems.iterator.runLengthEncode(parityOrdering).toSeq should be(expected)
+      elems.runLengthEncode(parityOrdering).toSeq should be(expected)
     }
 
     test("evens and odds") {
@@ -72,10 +72,9 @@ class RunLengthIteratorTest extends Suite {
 
   {
     test("re-encode") {
-      reencode(
-        Iterator('a' → 2, 'a' → 1, 'b' → 3, 'a' → 4, 'c' → 1, 'c' → 1, 'c' → 2, 'a' → 1)
-      )
-      .toList should be(
+      Iterator('a' → 2, 'a' → 1, 'b' → 3, 'a' → 4, 'c' → 1, 'c' → 1, 'c' → 2, 'a' → 1)
+        .reencode
+        .toList should be(
         List(
           'a' → 3,
           'b' → 3,
